feat: app export 경로 수정
