+++ src/app/layouts/AppLayout.tsx
@@ -1,0 +1,17 @@
+import Header from './Header.tsx'
+import Footer from './Footer.tsx'
+import PostsManagerPage from '@pages/PostsManagerPage'
+
+const App = () => {
+  return (
+    <div className="flex flex-col min-h-screen">
+      <Header />
+      <main className="flex-grow container mx-auto px-4 py-8">
+        <PostsManagerPage />
+      </main>
+      <Footer />
+    </div>
+  )
+}
+
+export default App
